requirements.txt
streamlit~=1.14.0
pandas~=1.4.3
black~=22.12.0
streamlit-authenticator~=0.2.1
streamlit-option-menu~=0.3.2
Pillow~=9.4.0
plotly~=5.11.0
numpy~=1.24.1
jinja2
pdfkit
python-dotenv